+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 2.3.5
- * All JC	Panther	mounts are now marked as flying
+ * Fixed certain JC	Panther	mounts were not marked as flying
  * Fixed unknown method of obtaining filters for ground mounts
 
 ## 2.3.4